Change messageReceived signatures
This way one can connect to the same signal either with or without
interest for the topic.

Another option had been to have two different messageReceived signals,
one including just the message and another the message plus the topic.
Then again, that would double the amount of signals and also cause
troubles connecting to a signal with different arguments, including
static cast magic.