Add conformance tests
Conformance tests are supposed to go through the specs and test each
item stated, to an extend possible. As a base, the paho test suite is
used, more specifically all tests of client_test.py.

In conjunction with tests for the Qt classes, we reach 72% conformance,
which is higher than the client test by paho reaches itself.

Change-Id: I5672a3664c340fe7f856de3eaf63bf117075acb6
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>