Make subscriptions active at request time
The paho conformance broker sends all retained messages first before
sending the SUBACK command. This contradicts to all existing solutions,
but is still valid. Section 3.8.4 of the standard says:
"The Server is permitted to start sending PUBLISH packets matching the
Subscription before the Server".

Hence, add the subscription already to the active list and allow
received messages to be emitted.

Change-Id: I1c90025ac561debc8ba4dc9957d1da2a153c8e67
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>