Fix connecting via an already open QIODevice transport
In case a transport is handed over to QMqttClient which is based on
QIODevice and already opened, the implementation was not sending a
CONNECT request and hence the broker never got informed about a new
client establishing connection. Hence manually send the connect
statement.

For QAbstractSocket based transports this was done already.

Task-number: QTBUG-66955
Change-Id: I50cba1b6dd03d0af561d4e39a3b57af5b2161a4b
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>