Disconnect when keepAlive fails to receive pingresp
The specs state that a client SHOULD disconnect after not receiving a
pingresp. So far, this needed to be checked manually. However, on Linux
a disrupted network connection does not imply QTcpSocket to switch to
disconnected state. Hence the client continues to consider the
connection active, leading to lost messages and incorrect state.

Task-number: QTBUG-74279
Change-Id: I6d8e540f7e89c621e5c010669085882d25440b0a
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>