Do not ignore MQTT 5 protocol violations in some more cases
The exceptional texts in 3.4.2.1, 3.5.2.1, 3.6.2.1, 3.7.2.1,
3.14.2.1 and 3.15.2.1 propose a micro optimization:
> The Reason Code and Property Length can be omitted
> if the Reason Code is 0x00 (Success) and there are no Properties.

which our current implementation treats like if it were
"The Property Length can be omitted if the Reason Code is 0x00".

In theory, this optimization could apply to CONNACK as well,
as there is a reason code and no payload, but that's not stated
anywhere in the spec and should not be assumed.

Change-Id: Ifaf9a3c45876661b8f7a4d688600f53b46288752
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>