Ensure error gets reported for the state in which it has been occurred
In certain scenarios, it is important to distinguish errors occurred
within particular states.
A common example is: if network error occurred for a connected
and operational client, one could try to reconnect and restore
previous MQTT session using the same clientId and cleanSession=false,
without having to bother the user if that succeуded;
however if network error occurred during connection to MQTT broker,
the behavior is completely different.

Change-Id: Id1e90b7916f3e2ab32320518d0da1100014e5dd1
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>