Update commercial license headers
Updated header.COMM to the files in tqtc-qtmqtt.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the the line count causes failures in tests.

Task-number: QTQAINFRA-4970
Change-Id: Ie1880e5f9ac487dc642b308b1a917395fa4eca0b
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>