QtMQTT: replace qSwap with std::swap/member-swap where possible
qSwap() is a monster that looks for ADL overloads of swap() and also
detects the noexcept of the wrapped swap() function, so it should only
be used when the argument type is unknown. In the vast majority of
cases, the type is known to be efficiently std::swap()able or to have
a member-swap. Call either of these.

As a drive-by, remove pointless inline adornments, port Q_DECL_NOTHROW
to noexcept, and fix the move-assignment operators to call member-swap
instead of swapping manually.

Task-number: QTBUG-97601
Change-Id: I2526194a426333e9f2773e5965458289459d1b4d
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
(cherry picked from commit 0cb10269bfbea3cfb02d340e47c7723562e240dc)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>