Reimplement QmlMqttClient
QMqttClient is not meant to be directly exported to QML. One reason is
that it uses quint16 for port declaration to minimize size of the class.
However, that is not properly handled on all platforms for the QML
engine.

Instead create a proper wrapper for the client and only expose required
elements.

Fixes: QTBUG-80440
Pick-to: 6.4
Change-Id: Id4066ce24338b0270af05b06c6216c1755dab533
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>