Examples: websocket IO device must be sequential
When using a websocket as the underlying IO device we must specify
that the device is sequential, to avoid spurious protocol violation
condition when packets are split across websocket frame boundaries.

The standard specifically states: "The receiver MUST NOT assume that
MQTT Control Packets are aligned on WebSocket frame boundaries."
New MQTT brokers like FlashMQ rely on well-behaved receivers.

Change-Id: Ia8e7fb0cf1cc518b364625b1955fb6b9ee44b7ef
Pick-to: 6.5
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>