Mark more files as security-critical
The qmqttclient.cpp file contains parts of the MQTT protocol
implementation, and hence marked as security-critical.

The qmqtttopicname.cpp file contains the code that validates the topic
name and also parses the levels -> it is a data-parser, so should be
security-critical as well.

Amends 18548033e15a4f5952065fe5ba6cd764b4e356a2, which incorrectly
marked these files as significant.

Not picking to 6.9 anymore, because that branch is already closed.

Task-number: QTBUG-135764
Pick-to: 6.10 6.8
Change-Id: I0e8b72fb1d24b7bbf6ac5edab63e137f9d36dd30
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>